+++ tests/test_inspect.py
@@ -7,14 +7,17 @@ from pytest_container.runtime import OciRuntimeBase
 
 from .test_container_build import LEAP
 
+_CTR_NAME = "foobar-12345"
 
 IMAGE_WITH_EVERYTHING = DerivedContainer(
+    singleton=True,
+    extra_launch_args=["--name", _CTR_NAME],
     base=LEAP,
     containerfile="""VOLUME /src/
 EXPOSE 8080 666
 RUN useradd opensuse
 USER opensuse
-ENTRYPOINT /bin/false
+ENTRYPOINT /bin/bash
 ENV HOME=/src/
 ENV MY_VAR=
 ENV SUFFIX_NAME=dc=example,dc=com
@@ -23,13 +26,18 @@ CMD ["/bin/sh"]
 )
 
 
-@pytest.mark.parametrize("container", [IMAGE_WITH_EVERYTHING], indirect=True)
-def test_inspect(container: ContainerData, container_runtime: OciRuntimeBase):
-    inspect = container.inspect
+@pytest.mark.parametrize(
+    "container_per_test", [IMAGE_WITH_EVERYTHING], indirect=True
+)
+def test_inspect(
+    container_per_test: ContainerData, container_runtime: OciRuntimeBase, host
+) -> None:
+    inspect = container_per_test.inspect
 
-    assert inspect.id == container.container_id
+    assert inspect.id == container_per_test.container_id
+    assert inspect.name == _CTR_NAME
     assert inspect.config.user == "opensuse"
-    assert inspect.config.entrypoint == ["/bin/sh", "-c", "/bin/false"]
+    assert inspect.config.entrypoint == ["/bin/sh", "-c", "/bin/bash"]
 
     assert (
         "HOME" in inspect.config.env and inspect.config.env["HOME"] == "/src/"
@@ -39,9 +47,9 @@ def test_inspect(container: ContainerData, container_runtime: OciRuntimeBase):
     # prefixes it with `localhost` and the full build tag
     # (i.e. `pytest_container:$digest`), while docker just uses the digest
     expected_img = (
-        str(container.container)
+        str(container_per_test.container)
         if container_runtime.runner_binary == "docker"
-        else f"localhost/pytest_container:{container.container}"
+        else f"localhost/pytest_container:{container_per_test.container}"
     )
 
     assert inspect.config.image == expected_img
@@ -58,4 +66,8 @@ def test_inspect(container: ContainerData, container_runtime: OciRuntimeBase):
         len(inspect.mounts) == 1
         and isinstance(inspect.mounts[0], VolumeMount)
         and inspect.mounts[0].destination == "/src"
+    )
+
+    assert inspect.network.ip_address or "" == host.check_output(
+        f'{container_runtime.runner_binary} inspect --format "{{{{ .NetworkSettings.IPAddress }}}}" {_CTR_NAME}'
     )